add insert, update, and delete on book service

+++ project_pbo/Models/BookModel.cs
@@ -3,7 +3,7 @@
     public class BookModel
     {
 
-        public Int16 BookId { get; set; }
+        public Int32 BookId { get; set; }
         public string? Title { get; set; }
         public string? Rating { get; set; }
         public string? Isbn { get; set; }
